move the enum adn type folder to the src folder
